# openapi-spec.json
{"openapi":"3.0.0","paths":{"/users/getAll":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"201":{"description":"Users fetched succesfully "}},"summary":"Get All Users","tags":["Users"]}},"/users/getById/{id}":{"get":{"operationId":"UsersController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/update/{id}":{"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/delete/{id}":{"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/me":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/updatePassword/{id}":{"patch":{"operationId":"UsersController_updatePassword","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/toggleStatus/{id}":{"patch":{"operationId":"UsersController_toggleStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/makeAdmin/{id}":{"patch":{"operationId":"UsersController_makeAdmin","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninDto"}}}},"responses":{"201":{"description":""}},"summary":"Google OAuth Login","tags":["Auth"]}},"/auth/google":{"get":{"operationId":"AuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}},"summary":"Initiate Google OAuth2 Login","tags":["Auth"]}},"/auth/google/redirect":{"get":{"operationId":"AuthController_googleAuthRedirect","parameters":[],"responses":{"200":{"description":""}},"summary":"Google OAuth2 Callback Handler","tags":["Auth"]}},"/auth/health":{"get":{"operationId":"AuthController_healthCheck","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/address/create":{"post":{"operationId":"AddressController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/getAll":{"get":{"operationId":"AddressController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/getById/{id}":{"get":{"operationId":"AddressController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/getByUser/{id}":{"get":{"operationId":"AddressController_getByUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/getmyaddress":{"get":{"operationId":"AddressController_getMyAddress","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/update/{id}":{"patch":{"operationId":"AddressController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAddressDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/setAsDefault/{id}":{"patch":{"operationId":"AddressController_setAsDefault","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/setAsBilling/{id}":{"patch":{"operationId":"AddressController_setAsBilling","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/address/delete/{id}":{"delete":{"operationId":"AddressController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Address"]}},"/categories/create":{"post":{"operationId":"CategoriesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["Categories"]}},"/categories/getAll":{"get":{"operationId":"CategoriesController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"query","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Categories"]}},"/categories/{id}":{"get":{"operationId":"CategoriesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Categories"]}},"/categories/update/{id}":{"patch":{"operationId":"CategoriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["Categories"]}},"/products/create":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["Products"]}},"/products/getAll":{"get":{"operationId":"ProductsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/getById/{id}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/findByCategory/{categoryId}":{"get":{"operationId":"ProductsController_findByCategory","parameters":[{"name":"categoryId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/update/{id}":{"patch":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/remove/{id}":{"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/toggleActivation/{id}":{"patch":{"operationId":"ProductsController_toggleActivation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/updatePrice/{id}":{"patch":{"operationId":"ProductsController_updatePrice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePriceDto"}}}},"responses":{"200":{"description":""}},"tags":["Products"]}},"/products/updateStock/{id}":{"patch":{"operationId":"ProductsController_updateStock","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStockDto"}}}},"responses":{"200":{"description":""}},"tags":["Products"]}},"/orders/create":{"post":{"operationId":"OrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Orders"]}},"/orders/getAll":{"get":{"operationId":"OrdersController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Orders"]}},"/order-items":{"post":{"operationId":"OrderItemsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderItemDto"}}}},"responses":{"201":{"description":""}},"tags":["OrderItems"]}},"/payments/initiate":{"post":{"operationId":"PaymentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentDto"}}}},"responses":{"201":{"description":""}},"tags":["Payment"]}},"/payments/history":{"get":{"operationId":"PaymentController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Payment"]}},"/payments/history/{id}":{"get":{"operationId":"PaymentController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Payment"]}},"/payments/webhook":{"post":{"operationId":"PaymentController_paystackWebhook","parameters":[],"responses":{"201":{"description":""}},"tags":["Payment"]}},"/payments/verify":{"post":{"operationId":"PaymentController_verifyPaymentWithPaystack","parameters":[],"responses":{"201":{"description":""}},"tags":["Payment"]}},"/payments/history/me":{"get":{"operationId":"PaymentController_getMyPaymentHistory","parameters":[],"responses":{"200":{"description":""}},"tags":["Payment"]}},"/payments/webhook/log":{"get":{"operationId":"PaymentController_getWebhookLog","parameters":[{"name":"page","required":true,"in":"path","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Payment"]}},"/product-images/create":{"post":{"operationId":"ProductImagesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductImageDto"}}}},"responses":{"201":{"description":""}},"tags":["ProductImages"]}},"/product-images/update/{id}":{"patch":{"operationId":"ProductImagesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductImageDto"}}}},"responses":{"200":{"description":""}},"tags":["ProductImages"]}},"/product-images/delete/{id}":{"delete":{"operationId":"ProductImagesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ProductImages"]}},"/product-images":{"get":{"operationId":"ProductImagesController_GetAll","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ProductImages"]}}},"info":{"title":"E-comm Api","description":"Description here","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/api/v1"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UpdateUserDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe","description":"The user name"},"email":{"type":"string","example":"john.doe@example.com","description":"The user email"},"password":{"type":"string","example":"password123","description":"The user password"}},"required":["name","email","password"]},"SigninDto":{"type":"object","properties":{}},"CreateAddressDto":{"type":"object","properties":{}},"UpdateAddressDto":{"type":"object","properties":{}},"CreateCategoryDto":{"type":"object","properties":{}},"UpdateCategoryDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{}},"UpdateProductDto":{"type":"object","properties":{}},"UpdatePriceDto":{"type":"object","properties":{}},"UpdateStockDto":{"type":"object","properties":{}},"OrderItemDto":{"type":"object","properties":{}},"CreateOrderDto":{"type":"object","properties":{"address_id":{"type":"string","example":"6f0e7f66-f8f4-45d9-89cf-2f9d3c6a4707"},"items":{"description":"List of items in the order","example":[{"product_id":"6f0e7f66-f8f4-45d9-89cf-2f9d3c6a4707","quantity":2}],"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}}},"required":["address_id","items"]},"CreatePaymentDto":{"type":"object","properties":{}},"CreateProductImageDto":{"type":"object","properties":{}},"UpdateProductImageDto":{"type":"object","properties":{}}}}}